Filtro
Otimização e limpeza de alguns aspectos do código
Criação da função extra de filtro.
