+++ src/App.js
@@ -3,31 +3,23 @@ import Map from './Map';
 import 'leaflet/dist/leaflet.css';
 import './App.css';
 import logo from './img/aiko.png';
-import Map2 from './Map2';
-import ReactDOM from 'react-dom';
-import { createRoot } from 'react-dom/client';
 import './Map.css';
-
+import Select from './Select';
 
 
 function App() {
     const mapaCompletoRef = useRef(null);
     
-    const MapButton = (equip, lista) => {
-      const x = lista.findIndex(obj => obj.key === equip);
-      createRoot(document.querySelector('.mapa-completo')).render(<Map2 Retorna={Retorna} equipmentId={equip} marker={lista[x]}/>);
-    };
-    const Retorna = () =>{
-      createRoot(document.querySelector('.mapa-completo')).render(<div><Map MapButton={MapButton}/></div>);
-    }
     return (
       
     <div className="App">
       <div className='logo'>
         <img src={logo} alt='logo' className='logoimg'/>
-      </div>
+        
+        <Select />
+      </div> 
       <div className='mapa-completo'  ref={mapaCompletoRef}>
-        <Map MapButton={MapButton}/>
+        <Map filter={null} />
       </div>
       <div className='triangulo'></div>
     </div>
